% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Confusion Matrix.R
\name{conf.matrix}
\alias{conf.matrix}
\title{Creates Confusion Matrix}
\usage{
conf.matrix(response.values, predicted.values)
}
\arguments{
\item{response.values}{Actual respose values}

\item{predicted.values}{Predicted respose values}
}
\value{
A list of accuracy, misclassification.rate, true.positive.rate, false.positive.rate, specificity, precision, prevalence
}
\description{
Takes in both actual and predicted values and calculates various metrcis
}
