% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CS Kurtosis.R
\name{cs_kurt}
\alias{cs_kurt}
\title{Creates CS Kurtosis}
\usage{
cs_kurt(data_set)
}
\arguments{
\item{data_set}{A data set used to calculate CS Kurtosis}
}
\value{
CS Kurtosis
}
\description{
Takes in a data set and calculates CS Kurtosis
}
