% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bootstrapped Bin Interval.R
\name{int.bin.boot}
\alias{int.bin.boot}
\title{Creates various metrics of the original data}
\usage{
int.bin.boot(n.boot = 750, PIC = 0.95, n = 10, mu = 8, sig = 2,
  b.CL = 0.999, inc = 0.01, seed = NULL, testCL = 0.95, m = 1,
  symmetric = TRUE, mean_boot_means = TRUE)
}
\arguments{
\item{n.boot}{Number of bootstrap samples you want to use (# of intervals)}

\item{PIC}{The minimum percentage of boostrapped CI's that must lie within prediction region}

\item{n}{Sample size of simulated data}

\item{mu}{Mean of simulated data}

\item{sig}{Standard deviation of simulated data}

\item{b.CL}{Confidence level used for bootstrap CI}

\item{inc}{Size increments (Construction of prediction region)}

\item{seed}{Avoids randomizations according to the seed placed}

\item{testCL}{Use this to be able to test any CL for t-test against that of the prediction region}

\item{m}{Used to determine what iteration the data was created; only used for plotting purposes}

\item{symmetric}{If TRUE, a symmetric prediction region will be created; else, an unsymmetric prediction region will be created}

\item{mean_boot_means}{If TRUE, the mean of bootstrapped means will be used as the starting point; else, the mean of the original data will be used}
}
\value{
Various metrics based of off the simulated data
}
\description{
Generates normal data set according to parameters, builds multiple bootstrap intervals & calculates various normality metrics. Also generates a visual of the bootstrapped intervals along with the prediction region and t-interval
}
