% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K Folds.R
\name{kfold}
\alias{kfold}
\title{Runs k-folds on either rpart or ctree}
\usage{
kfold(ctree.OR.rpart, data_frame, cp.min = 0, k = 5, hseed = 10)
}
\arguments{
\item{ctree.OR.rpart}{User specified to use rpart or ctree}

\item{data_frame}{Data to be used}

\item{cp.min}{Minimum cp value to be used if rpart is selected}

\item{k}{Number of partitions in which to partition the data}

\item{hseed}{Avoids randomizations according to the seed placed}
}
\value{
Metrics from confusion matrix
}
\description{
Depending on the call (rpart or ctree), runs either with data split into k partitions
}
