% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lap.R
\name{lap}
\alias{lap}
\title{Creates plots from multiple runs (i.e 1 lap) along with confusion matrix metrics from both rpart & ctree}
\usage{
lap(reps = 725, sample.seed = 9687, int.seed = 10000, n.boot.lap = 750,
  PIC.lap = 0.95, n.lap = 10, mu.lap = 8, sig.lap = 2, inc.lap = 0.01,
  b.CL.lap = 0.999, testCL.lap = 0.95, symmetric.lap = TRUE,
  mean_boot_means.lap = TRUE)
}
\arguments{
\item{reps}{Number of prediction regions to test to create a single decision tree}

\item{sample.seed}{Initial seed for "random" future seeds}

\item{int.seed}{Largest possoble value to sample to obtain random seeds for every run in the lap}

\item{n.boot.lap}{Number of bootstrap samples you want to use (# of intervals)}

\item{PIC.lap}{The minimum percentage of boostrapped CI's that must lie within prediction region}

\item{n.lap}{Sample size of simulated data}

\item{mu.lap}{Mean of simulated data}

\item{sig.lap}{Standard deviation of simulated data}

\item{inc.lap}{Size increments (Construction of prediction region)}

\item{b.CL.lap}{Confidence level used for bootstrap CI}

\item{testCL.lap}{Use this to be able to test any CL for t-test against that of the prediction region}

\item{symmetric.lap}{If TRUE, a symmetric prediction region will be created; else, an unsymmetric prediction region will be created}

\item{mean_boot_mean.lap}{If TRUE, the mean of bootstrapped means will be used as the starting point; else, the mean of the original data will be used}
}
\value{
A list with metrics from confusion matrix from both rpart and ctree
}
\description{
Uses multiple runs (reps; i.e 1 lap) to create decision tree to improve empirical CL for prediction region & plots normal metrics
}
