% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Laps Ensemble.R
\name{laps_ensemble}
\alias{laps_ensemble}
\title{Creates ensemble of trees}
\usage{
laps_ensemble(emp.reps = 701, reps = 725, int.seed = 10000,
  n.boot.ens = 750, PIC.ens = 0.95, n.ens = 10, mu.ens = 8,
  sig.ens = 2, inc.ens = 0.01, b.CL.ens = 0.999, testCL.ens = 0.95,
  symmetric.ens = TRUE, mean_boot_means.ens = TRUE)
}
\arguments{
\item{emp.reps}{Number of decision trees (i.e laps)}

\item{reps}{Number of prediction regions to test to create a single decision tree}

\item{int.seed}{Largest possoble value to sample to obtain random seeds for every lap}

\item{n.boot.ens}{Number of bootstrap samples you want to use (# of intervals)}

\item{PIC.ens}{The minimum percentage of boostrapped CI's that must lie within prediction region}

\item{n.ens}{Sample size of simulated data}

\item{mu.ens}{Mean of simulated data}

\item{sig.ens}{Standard deviation of simulated data}

\item{inc.ens}{Size increments (Construction of prediction region)}

\item{b.CL.ens}{Confidence level used for bootstrap CI}

\item{testCL.ens}{Use this to be able to test any CL for t-test against that of the prediction region}

\item{symmetric.ens}{If TRUE, a symmetric prediction region will be created; else, an unsymmetric prediction region will be created}

\item{mean_boot_mean.ens}{If TRUE, the mean of bootstrapped means will be used as the starting point; else, the mean of the original data will be used}
}
\value{
A list composed of rpart models, ctree models, rpart confusion matrices, ctree confusion matrices & empirical quantities for t-test, bootstrap percentile & prediction region intervals. Also creates a pdf in working directorate of every lap.
}
\description{
Uses multiple laps to create ensemble of trees to create vote as to whether this method works for a specified data set
}
