% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mean Absolute Deviation.R
\name{mean_abs_dev}
\alias{mean_abs_dev}
\title{Creates Mean Absolute Difference}
\usage{
mean_abs_dev(df)
}
\arguments{
\item{df}{A data set used to calculate the mean absolute difference}
}
\value{
Mean Absolute Difference
}
\description{
Takes in a data set and calculates the Mean Absolute Difference
}
