% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Real Bootstrapped Bin Interval.R
\name{real.int.bin.boot}
\alias{real.int.bin.boot}
\title{Creates prediction region for real data}
\usage{
real.int.bin.boot(df, n.boot = 750, PIC = 0.95, b.CL = 0.999,
  inc = 0.01, seed = NULL, testCL = 0.95, m = 1, symmetric = TRUE,
  mean_boot_means = TRUE)
}
\arguments{
\item{df}{Data to be used to construct prediction region}

\item{n.boot}{Number of bootstrap samples you want to use (# of intervals)}

\item{PIC}{The minimum percentage of boostrapped CI's that must lie within prediction region}

\item{b.CL}{Confidence level used for bootstrap CI}

\item{inc}{Size increments (Construction of prediction region)}

\item{seed}{Avoids randomizations according to the seed placed}

\item{testCL}{Use this to be able to test any CL for t-test against that of the prediction region}

\item{m}{Used to determine what iteration the data was created; only used for plotting purposes}

\item{symmetric}{If TRUE, a symmetric prediction region will be created; else, an unsymmetric prediction region will be created}

\item{mean_boot_means}{If TRUE, the mean of bootstrapped means will be used as the starting point; else, the mean of the original data will be used}
}
\value{
Gives prediction region endpoints along with a plot
}
\description{
Creates prediction region aling with t-interval & bootstrap percentile interval
}
