% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Round.R
\name{round.func}
\alias{round.func}
\title{Rounds differently than 0.5 cutoff}
\usage{
\method{round}{func}(x, r.val)
}
\arguments{
\item{x}{A list of numbers between 0 and 1 to be rounded}

\item{r.val}{Cutoff value to round up}
}
\value{
a vector of rounded values
}
\description{
Rounds a list of numbers between 0 & 1 based on specified cutoff
}
