% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data Splitter.R
\name{split_data.header}
\alias{split_data.header}
\title{Splits the data into k partitions}
\usage{
split_data.header(data.set, n.folds, seed = NULL)
}
\arguments{
\item{data.set}{Data to be used}

\item{n.folds}{Number of partitions of the data}

\item{seed}{Avoids randomizations according to the seed placed}
}
\value{
The data randomly split into k partitions
}
\description{
Randomizes the data and then partitions it out into k folds
}
